+++ src/test/java/com/ticket/captain/festivalDetail/FestivalDetailRestControllerTest.java
@@ -19,6 +19,7 @@ import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.security.test.context.support.WithUserDetails;
 import org.springframework.test.web.servlet.MockMvc;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
@@ -65,7 +66,7 @@ public class FestivalDetailRestControllerTest {
         FestivalDetailCreateDto festivalDetailCreateDto = FestivalDetailCreateDto.builder()
                 .salesType(SalesType.FREE.toString())
                 .amount(10000L)
-                .price(20000L)
+                .price(BigDecimal.valueOf(20000))
                 .drawDate(LocalDateTime.now())
                 .processDate(LocalDateTime.now())
                 .build();
@@ -81,7 +82,7 @@ public class FestivalDetailRestControllerTest {
         FestivalDetailCreateDto festivalDetailCreateDto = FestivalDetailCreateDto.builder()
                 .salesType(SalesType.FREE.toString())
                 .amount(1000L)
-                .price(2000L)
+                .price(BigDecimal.valueOf(2000))
                 .drawDate(LocalDateTime.now())
                 .processDate(LocalDateTime.now())
                 .build();
@@ -120,7 +121,7 @@ public class FestivalDetailRestControllerTest {
         //given
         FestivalDetailUpdateDto festivalDetailUpdateDto = FestivalDetailUpdateDto.builder()
                 .amount(300L)
-                .price(500L)
+                .price(BigDecimal.valueOf(500L))
                 .drawDate(LocalDateTime.now())
                 .processDate(LocalDateTime.now())
                 .salesType(SalesType.DRAW.toString())
